feat: scale/offset/unit for setpoint commands, engineering-value entry
Setpoint (promptForValue) commands can now be configured with a scale
factor, offset, and unit — the same conversion a data point already
uses to decode a reading, applied in reverse for writing one. An
operator now types a real engineering value (e.g. 30 seconds, 22.5°C)
instead of having to already know the PLC's raw register encoding;
DeviceService.issueCommand converts it before dispatch:

    raw = round((entered - offset) / scaleFactor)

- CommandTemplate gains nullable scaleFactor/offset/unit columns
  (default to 1/0/no-op when absent, so existing commands are
  unaffected). scaleFactor of zero is rejected at save time.
- IssueCommandRequest.value is now a BigDecimal (was Integer), so
  decimal engineering values can be sent at all.
- The existing single-register range check (-32768..65535) now runs
  against the *converted* raw value, catching a badly-scaled or
  mistyped entry the same way it already caught raw ones.
- Value entry / edit forms gain a Scale / Offset / Unit row, styled
  like the identical fields on the Data Point form. The send dialog
  shows the configured unit next to the Value field.

Verified end-to-end via direct API calls and the real form: entering
30 with scale=10 sends raw 3; entering 22.5 with scale=0.1/offset=-40
sends raw 625; a zero scale factor and an out-of-range scaled value
are both rejected with clear 400s.

+++ ecoFarm/src/main/java/com/ecoFarm/api/v1/dto/request/IssueCommandRequest.java
@@ -2,11 +2,15 @@ package com.ecoFarm.api.v1.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record IssueCommandRequest(
     @NotNull UUID commandTemplateId,
     /** Required when the template has promptForValue set; ignored otherwise —
-     * fixed commands always send their own stored value, never a client-supplied one. */
-    Integer value
+     * fixed commands always send their own stored value, never a client-supplied one.
+     * This is the operator-facing engineering value (e.g. 22.5 for °C, 30 for
+     * seconds) — DeviceService converts it to a raw register value using the
+     * command's own scaleFactor/offset before dispatch. */
+    BigDecimal value
 ) {}
